requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
earthengine-api==0.1.382
google-auth==2.23.4
google-cloud-storage==2.10.0
python-dotenv==1.0.0
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
numpy>=1.24.4
pillow==10.1.0
httpx==0.25.2
pyproj==3.6.1
psutil==5.9.6
